Move new models to correct package
